LayoutTests/http/wpt/html/browsers/the-window-object/window-open-noopener-webkit.html

<!doctype html>
<meta charset=utf-8>
<title>window.open() with "noreferrer" tests</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
async_test(t => {
  var uniqueTestId = Math.floor((Math.random() * 1000000000000));
  w = window.open("support/noreferrer-target-webkit.html?" + uniqueTestId, "", "noreferrer");
  assert_equals(w, null);
  handle = setInterval(() => {
    if (localStorage.getItem(uniqueTestId + "-name") === null)
        return;

    clearInterval(handle);
    assert_equals(localStorage.getItem(uniqueTestId + "-name"), "");
    assert_equals(localStorage.getItem(uniqueTestId + "-referrer"), "");
    assert_equals(localStorage.getItem(uniqueTestId + "-hasOpener"), "false");

    t.done();
  }, 10);
});
</script>
