LayoutTests/platform/wpe/http/wpt/html/browsers/the-window-object/window-open-noopener-webkit-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.

Harness Error (TIMEOUT), message = null

TIMEOUT window.open() with "noreferrer" tests Test timed out
